data_ingestion/requirements.txt
langchain
chromadb
openai
python-dotenv
sentence-transformers
numpy
langchain-community